+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4100,7 +4100,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KYVW8YY6BTN6WPVSWARM344S"),
+  id: z.string().default("01KYW65M9YZ23FRARFBQDPFNJQ"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),