+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-31T10:42:00.647Z",
+    "2026-07-31T13:34:57.400Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {