+++ packages/orq-rc/RELEASES.md
@@ -11329,3 +11329,13 @@ Based on:
 - [typescript v4.13.0-rc.24] packages/orq-rc
 ### Releases
 - [NPM v4.13.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.24 - packages/orq-rc
+
+## 2026-07-31 13:38:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.13.0-rc.25] packages/orq-rc
+### Releases
+- [NPM v4.13.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.25 - packages/orq-rc